% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{identifyBetaPars}
\alias{identifyBetaPars}
\title{Determine the best-fit beta distribution for a given confidence interval for a probability parameter.}
\usage{
identifyBetaPars(pLow, pUpp, alpha = 0.05, initPars = c(50, 50))
}
\arguments{
\item{pLow}{The observed lower quantile .}

\item{pUpp}{The observed upper quantile.}

\item{alpha}{The confidence level; i.e. the desired coverage is 1-alpha. Defaults to 0.05.}

\item{initPars}{A vector of length 2 giving the initial parameter values to start the optimisation; defaults to c(50,50).}
}
\value{
A vector of length 2 giving the 2 parameter shape1 and shape1 for use with rebta/dbeta/ppbeta/qbeta
}
\description{
Finds the best-fit beta distribution for a given confidence interval for a probability parameter and returns the shape1
}
\examples{
identifyBetaPars(pLow=0.1167,pUpp=0.1636,initPars=c(200,800))

}
\seealso{
\code{\link{ssBetaPars}}, \code{\link{optim}}, \code{\link{dbeta}}
}
